JS: Big refactor to do less work when turbolinking to new page
We were creating a lot of event handlers on every page transition. Nearly all of
them could be delegated to `document`, which lets us reuse the same handlers on
every page in a given turbolinks session. Which, come to think of it,
re-using global behavior instead of reinstantiating it every time you navigate
was supposed to be the main benefit of turbolinks in the first place.